Updated autocomplete to handle upper and lower case lettes
